+++ gtfs_kit/stops.py
@@ -648,26 +648,23 @@ def build_geometry_by_stop(
 def stops_to_geojson(feed: "Feed", stop_ids: Optional[Iterable[str]] = None) -> dict:
     """
     Return a GeoJSON FeatureCollection of Point features
-    representing ``feed.stops``.
+    representing all the stops in ``feed.stops``.
     The coordinates reference system is the default one for GeoJSON,
     namely WGS84.
 
     If an iterable of stop IDs is given, then subset to those stops.
-    If the subset is empty, then return a FeatureCollection with an empty list of
-    features.
     If some of the given stop IDs are not found in the feed, then raise a ValueError.
     """
-    if stop_ids is not None:
-        D = set(stop_ids) - set(feed.stops.stop_id)
-        if D:
-            raise ValueError(f"Stops {D} are not found in feed.")
+    if stop_ids is None or not list(stop_ids):
+        stop_ids = feed.stops.stop_id
+
+    D = set(stop_ids) - set(feed.stops.stop_id)
+    if D:
+        raise ValueError(f"Stops {D} are not found in feed.")
 
     g = geometrize_stops(feed, stop_ids=stop_ids)
-    if g.empty:
-        result = {"type": "FeatureCollection", "features": []}
-    else:
-        result = hp.drop_feature_ids(json.loads(g.to_json()))
-    return result
+
+    return hp.drop_feature_ids(json.loads(g.to_json()))
 
 
 def get_stops_in_area(