+++ tests/test_trips.py
@@ -106,6 +106,13 @@ def test_compute_trip_stats():
     )
     assert set(trip_stats.columns) == expect_cols
 
+    # Distance units should be correct
+    d1 = trip_stats.distance.iat[0]  # km
+    trip_stats_2 = compute_trip_stats(feed.convert_dist("ft"), route_ids=rids)
+    d2 = trip_stats_2.distance.iat[0]  # mi
+    f = get_convert_dist("km", "mi")
+    assert f(d1) == d2
+
     # Shapeless feeds should have null entries for distance column
     feed = cairns_shapeless.copy()
     trip_stats = compute_trip_stats(feed, route_ids=rids)
@@ -189,7 +196,11 @@ def test_trips_to_geojson():
     assert len(gj["features"]) == n
 
     gj = trips_to_geojson(feed, trip_ids, include_stops=True)
-    k = feed.stop_times.loc[lambda x: x.trip_id.isin(trip_ids), "stop_id"].nunique()
+    k = (
+        feed.stop_times.loc[lambda x: x.trip_id.isin(trip_ids)]
+        .drop_duplicates(subset=["trip_id", "stop_id"])
+        .shape[0]
+    )
     assert len(gj["features"]) == n + k
 
     with pytest.raises(ValueError):